refactor(cursor-based-pagination): separate data and UI concerns in Boundary architecture
- move SuspenseCursorVerticalInfiniteScrollBoundary → SuspenseCursorPaginationBoundary
- move CursorVerticalInfiniteScrollBoundary → CursorPaginationBoundary
- relocate Boundary components to /src/shared/ui/Boundary/
- keep Container components in /src/shared/ui/InfiniteScroll/
- update InfiniteScroll index to reflect new architecture
- improve naming consistency: focus on pagination logic vs UI rendering

Related to: #82
